Separate installer components from installer app container
