+++ src/renderer/src/components/installer/index.ts
@@ -1,0 +1,2 @@
+export * from "./installer-styled";
+export * from "./installer-content";
